Remove v2 views in regcore
