+++ solution/backend/regcore/urls.py
@@ -1,18 +1,7 @@
 from django.urls import path, include
 
+from regcore.admin import BulkSynonymView
 from regcore.views import (
-    EffectivePartView,
-    EffectiveTitlesView,
-    EffectivePartsView,
-    EffectivePartTocView,
-    PartsView,
-    ExistingPartsView,
-    ParserConfigurationView,
-    PartListView,
-    BulkSynonymView
-)
-
-from regcore.v3views import (
     title,
     part,
     contents,
@@ -25,19 +14,6 @@ from regcore.v3views import (
 
 urlpatterns = [
     path("admin/bulk_synonyms", BulkSynonymView.as_view(), name="bulk_synonyms"),
-    path("v2/", include([
-        path("", include('regcore.search.urls')),
-        path("", include('resources.urls')),
-        path("", PartsView.as_view(), name="regcore"),
-        path("parser_config", ParserConfigurationView.as_view()),
-        path("all_parts", PartListView.as_view()),
-        path("title/<title>/part/<name>", PartsView.as_view()),
-        path("<date>", EffectiveTitlesView.as_view()),
-        path("<date>/title/<title>", EffectivePartsView.as_view()),
-        path("<date>/title/<title>/part/<name>", EffectivePartView.as_view()),
-        path("<date>/title/<title>/part/<name>/toc", EffectivePartTocView.as_view()),
-        path("title/<title>/existing", ExistingPartsView.as_view()),
-    ])),
     path("v3/", include([
         path("resources/", include('resources.v3urls')),
         path("", include('regcore.search.v3urls')),
