+++ solution/backend/regcore/views/contents.py
