+++ solution/backend/regcore/views/metadata.py
