+++ solution/backend/regcore/views/mixins.py
