+++ solution/backend/regcore/views/part.py
