+++ solution/backend/regcore/views/synonyms.py
