+++ solution/backend/regcore/views/utils.py
